reorganise Python files into src dir
