+++ setup.cfg
@@ -2,7 +2,7 @@
 testpaths = tests
 
 [coverage:run]
-source = set_creds
+source = src
 
 [coverage:html]
 title = Code Coverage Report
